requirements.txt
streamlit
r2r
pandas
numpy
nest_asyncio
requests
asyncio
aiohttp
deprecated
google-auth-oauthlib
google-api-python-client
python-jwt
python-dotenv
extra-streamlit-components
supabase
